requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pillow>=8.2.0
opencv-python>=4.5.0
opencv-python-headless>=1.0.0
albumentations>=1.0.0
tqdm>=4.60.0
scikit-learn>=0.24.0
torch>=1.9.0
torchvision>=0.10.0
torch-geometric>=2.0.0
networkx>=2.5.0
wandb>=0.15.7
